route adjustments and implementations

+++ ExpenseTracker.Application/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace ExpenseTracker.Application.Interfaces
 {
     public interface IUserRepository
     {
-        Task<User?> GetById(int id);
+        Task<User?> GetById(long id);
         Task<User?> GetByUsername(string username);
         Task<User?> CreateUser(User user);
     }
